requirements.txt
streamlit==1.31.0
streamlit-elements==0.1.0
streamlit-aggrid==0.3.4
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25
pyyaml==6.0.1
plotly==5.18.0
altair==5.2.0
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3